Check if the build already exists before requesting build

+++ src/services/BuildRequester.groovy
@@ -7,7 +7,9 @@ import main.util.GithubHelper
 import main.util.ProcessRunner
 import main.exception.TravisHelperException
 import main.util.FileManager
+import main.util.GithubHelper
 import static main.app.MiningFramework.arguments
+
 
 import java.text.SimpleDateFormat
 
@@ -26,25 +28,51 @@ class BuildRequester implements DataCollector {
 
     public void collectData(Project project, MergeCommit mergeCommit) {
         if (arguments.providedAccessKey()) {
-            String branchName = mergeCommit.getSHA().take(5) + "_build_branch_${getCurrentTimestamp()}"
-            
-            checkoutCommitAndCreateBranch(project, branchName, mergeCommit.getSHA()).waitFor()
-            
-            File travisFile = new File("${project.getPath()}/.travis.yml")
             String[] ownerAndName = getRemoteProjectOwnerAndName(project)
-            travisFile.delete()
-            BuildSystem buildSystem = getBuildSystem(project)
+            String projectOwner = ownerAndName[0]
+            String projectName = ownerAndName[1]
 
-            if (buildSystem != BuildSystem.None) {
-                travisFile << getNewTravisFile(mergeCommit.getSHA(), ownerAndName[0], ownerAndName[1], buildSystem)
-                commitChanges(project, "'Trigger build #${mergeCommit.getSHA()}'").waitFor()
-                pushBranch(project, branchName).waitFor()
+            if (!buildAlreadyExists(projectOwner, projectName, mergeCommit)) {
+                String branchName = generateBranchName(mergeCommit)
+
+                checkoutCommitAndCreateBranch(project, branchName, mergeCommit.getSHA()).waitFor()
                 
-                goBackToMaster(project).waitFor()
-                println "${project.getName()} - Build requesting finished!"
-            }
+                File travisFile = getTravisFile(project)
+                travisFile.delete()
+                BuildSystem buildSystem = getBuildSystem(project)
 
+                if (buildSystem != BuildSystem.None) {
+                    travisFile << getNewTravisFile(mergeCommit.getSHA(), ownerAndName[0], ownerAndName[1], buildSystem)
+                    commitChanges(project, "'Trigger build #${mergeCommit.getSHA()}'").waitFor()
+                    pushBranch(project, branchName).waitFor()
+                    
+                    goBackToMaster(project).waitFor()
+                    println "${project.getName()} - Build requesting finished!"
+                }                
+            } else {
+                println "${project.getName()} - Build requesting skiped: build already exists"
+            }
+        } else {
+            println "${project.getName()} - Build requesting skiped: access key not provideds"
         }
+    }
+
+    private File getTravisFile(Project project) {
+        return new File("${project.getPath()}/.travis.yml")
+    }
+
+    private String generateBranchName(MergeCommit mergeCommit) {
+        return mergeCommit.getSHA().take(5) + "_build_branch_${getCurrentTimestamp()}"
+    }
+
+    private boolean buildAlreadyExists(String projectOwner, String projectName, MergeCommit mergeCommit) {
+        GithubHelper githubHelper = new GithubHelper(arguments.getAccessKey())
+
+        def releases = githubHelper.getRepositoryReleases(projectOwner, projectName);
+
+        def mergeCommitRelease = releases.find { release -> release.name.endsWith(mergeCommit.getSHA()) }
+
+        return mergeCommitRelease != null
     }
 
     private BuildSystem getBuildSystem (Project project) {
@@ -66,29 +94,29 @@ class BuildRequester implements DataCollector {
     }
 
     static private String[] getRemoteProjectOwnerAndName(Project project) {
-        String remoteUrl = ProcessRunner
-            .runProcess(project.getPath(), 'git', 'config', '--get', 'remote.origin.url').getText()
+        String remoteUrl = getRemoteUrl(project)
         String[] splitedValues = remoteUrl.split('/')
         return [splitedValues[splitedValues.size() - 2], splitedValues[splitedValues.size() - 1]]
     }
+
+    static private String getRemoteUrl(Project project) {
+        String remoteUrl = ProcessRunner.runProcess(project.getPath(), 'git', 'config', '--get', 'remote.origin.url').getText()
+
+        return remoteUrl.trim()
+    } 
 
     static private Process goBackToMaster(Project project) {
         return ProcessRunner.runProcess(project.getPath(), 'git', 'checkout', 'master')
     }
 
     static private Process pushBranch(Project project, String branchName) {
-        return ProcessRunner.runProcess(project.getPath(), 'git', 'push','-u', 'origin', branchName)
+        return ProcessRunner.runProcess(project.getPath(), 'git', 'push', 'origin', branchName)
     }
 
     static private Process commitChanges(Project project, String message) {
         ProcessRunner.runProcess(project.getPath(), "git", "add", ".travis.yml").waitFor()
 
         return ProcessRunner.runProcess(project.getPath(), "git", "commit", "-a", "-m", "${message}")
-    }
-
-    static private String getRemoteUrl(Project project) {
-        return ProcessRunner.
-            runProcess(project.getPath(), "git", "config", "--get", "remote.origin.url").getText()
     }
 
     static private String getBranchName(MergeCommit mergeCommit) {
@@ -127,8 +155,7 @@ before_deploy:
     - tar -zcvf result.tar.gz *
 deploy:
   provider: releases
-  api_key:
-    secure: \$GITHUB_TOKEN
+  api_key: \$GITHUB_TOKEN
   file: result.tar.gz
   name: fetchjar-${commitSha}
   file_glob: true
